x11/sddm: Add proper FreeBSD VirtualTerminal support
This is needed to support starting Wayland sessions, so that the running
Xorg gets notified about its VT being released and thus releases its DRM
and input devices, allowing the Wayland compositor to use them. Upstream
has some patches for VirtualTerminal_FreeBSD.cpp to make it not a stub
but they are incomplete and buggy, so better to just lightly adapt the
working Linux version and get something that's actually functional.

Note that the code to leak the VT FD into the Wayland session is needed
to support VT switching, as if nothing has it open then you can't switch
back to the VT once you switch away.