sysutils/u-boot-bhyve-arm64: New port
This is a port of https://github.com/zxombie/u-boot/tree/bhyvearm64 to
the version of u-boot in the ports tree.  There is a bhyve arm64 port
that is close to landing, so now seems like a good time to make the port
available.

In the longer term we would upstream this platform in some form, but
since arm64 bhyve is experimental, I think trying to upstream at this
point is premature.  Of course, we prefer not to carry patches in the
ports tree when possible, but in this case it seems justified to do so
for some time, to give users a change to test and report problems.

Approved by:	manu
Sponsored by:	Innovate UK
Sponsored by:	Klara, Inc. (hardware)
Differential Revision:	https://reviews.freebsd.org/D42437

(cherry picked from commit 2759a169bd8b3ecf209b9f7309a2c4e310918d86)