devel/llvm-base: depend on the absolute clang path
Mk/Scripts/actual-package-depends.sh removes a dependency on clang if it
is found in $PATH and does not belong to any package.

Since FreeBSD does include clang in the base system and hence it would
find it in $PATH as well, change the dependency on clang to use its
absolute path rather than a binary name.