Replace gnulib stdint.h with CHERI aware version
If a port contains stdint.in.h from gnulib and it does not contain
fixes for CHERI, replace it with an upstream version that contains the
required fixes.  We no longer need to modify it for CHERI, simply update
it to a newer version.

This version corresponds to gnulib commit
6e4aa95cc6f165d2d3a30e963f13596e6b09ed90.

Fixes: #129