sysutils/u-boot-bhyve-arm64: Enable MBR and ISO9660 partition support
These are both required to be supported for partition discovery by the
UEFI specification. Most arm64 media uses GPT so effectively doesn't
need this, but Debian's installer is an ISO with an additional MBR on
arm64 (though curiously on amd64 it's GPT+MBR on the ISO).

Note: Current Debian install media has a version of shim that has a bug
further preventing it from working with recent versions of U-Boot, but
once that is worked around it boots and installs.

(partially cherry picked from commit ba5777140ec41aae909456f6da77e1b336f52fcb)