Disable gobject-introspection for purecap ABI.
gobject-introspection is dependent of support in libffi for ffi
closures, which is current missing for the the purecap ABI. Although the
depet push origindency on g-ir-scanner is build time dependency it requires ABI
compatability so it is not possible to use the hybrid
gobject-introspection package. Therefore, support for introspection is
disabled.

As introspection is used solely for generating documentation and the
purecap ABI disables the DOCS option by default, disabling
gobject-introspection has no impact at present.  However, it is somewhat
disruptive as it requires modifications to the pkg-plist of various
ports.

Disable gobject-introspection for gtk30 deps.

librsvg2 requires building with the VAPI (Vala API) option unset. The
Vala API is generated with gobject-introspection which isn't currently
supported for purecap (due to its dependency on ffi closure support in
libffi).