Disable gobject-introspection for purecap ABI.
gobject-introspection is dependent of support in libffi
for ffi cloisures, which is current missing for the the
purecap ABI. Although the dependency on g-ir-scanner is
build time dependency it requires ABI compatability so
it is not possible to use the hybrid gobject-introspection
package. Therefore, support for introspection is disbaled.

As introspection is used solely for generating documentation
and the purecap ABI disbales the DOCS option by default,
disabling gobject-introspection has no impact at present.
However, it is somewhat disruptive as it requires modifications
to the ports pkg-plist.