emulators/qemu-cheri: package firmwares and other data
Add data (firmware binaries, keymaps, etc.) necessary to run
CheriBSD/Morello under QEMU on CheriBSD/Morello to the package. The data
are installed in $LOCALBASE/share/qemu-cheri not to conflict with other
QEMU ports.

Note that we currently don't have a port with BBL for CHERI-RISC-V.
Currently, users willing to use qemu-cheri with CheriBSD/CHERI-RISC-V
must build bbl-riscv64cheri-virt-fw_jump.bin with cheribuild. We should
add a port for this firmware as well.