Mk/Uses/python.mk: don't depend on directories
Use files from packages to register dependencies. Otherwise, the script
Mk/Scripts/actual-package-depends.sh executing pkg which will not be
able to find a dependency package based on a directory that was created
when a new file was stored in it for the dependency package. In such
a case, the dependency is not included in a package manifest.

While it's not relevant for CheriABI packages as we cannot register
dependencies between packages for different ABIs, the same logic would
fail when building hybrid ABI packages. In result, each time we rebuild
packages, Poudriere would discover that an existing package is missing
a dependency, it would remove broken packages with their dependant
packages and rebuild listed packages with their removed dependencies.

(cherry picked from commit 52b63857c0e6ccf48adef3ce69ddac25e648ce8f)