www/nginx: Hijack nginx port to build nginx-cheri.
Use our own github sources for nginx-cheri. This requires an extra
minor change to the nginx Makefile, but we should have better control
over what is being built. In particular, the nginx port has some extra
patches for HTTPv3 support that are not part of upstream nginx, and
are not ported to CHERI.