graphics/cairo: updated patch for CheriABI
The updated patch changes the use of types in the template file
from `gsize` to `GType`. This ensures capability provenance for
`gobject` type identifiers. As well as ensuring correct alignment for
allocation from a pool allocator.s

Increment the PORTREVISION as a result of the fixes.

Updated patch for graphics/cairo generated from [1].

[1] https://github.com/CTSRD-CHERI/cairo/tree/8a00e6f765793459d00238f3fff57969154561f6