www/chromium: import a patch to build with LLVM >=15
`__is_cpp17_contiguous_iterator` is not available in the current CHERI
LLVM for Morello version. Upstream has a workaround that defines the
symbol not to change the existing use cases of it.

Unfortunately, the commit with above workaround does not cleanly apply
to our Chromium version and I had to adjust it to the version.